+++ lib/edits_types/edit_type_enum.dart
@@ -1,0 +1,1 @@
+enum EditType { text, list, custom, uneditable }