+++ lib/edits_types/list_edit_tile.dart
@@ -1,0 +1,13 @@
+import 'dart:io';
+
+import 'package:flutter/material.dart';
+
+class ListEditTile<T> {
+  final Widget title;
+  final T value;
+
+  ListEditTile({
+    required this.title,
+    required this.value,
+  });
+}