+++ lib/platform_settings_ui.dart
@@ -5,3 +5,6 @@ export 'settings_switch_tile/settings_switch_tile.dart';
 export 'settings_cursor_tile/settings_slider_tile.dart';
 export 'settings_section/settings_section.dart';
 export 'settings_list/settings_list.dart';
+export 'edits_types/list_edit_tile.dart';
+export 'edits_types/list_edit_type_view.dart';
+export 'edits_types/edit_type_enum.dart';