+++ lib/settings_section/settings_section_ios.dart
@@ -1,5 +1,4 @@
 import 'package:flutter/material.dart';
-import 'package:platform_settings_ui/platform_settings_ui.dart';
 
 @protected
 class SettingsSectionIos extends StatelessWidget {
@@ -14,13 +13,22 @@ class SettingsSectionIos extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Padding(
-      padding: const EdgeInsets.symmetric(horizontal: 12.0, vertical: 8.0),
+      padding: const EdgeInsets.symmetric(horizontal: 12.0, vertical: 12.0),
       child: Column(
         mainAxisAlignment: MainAxisAlignment.start,
         crossAxisAlignment: CrossAxisAlignment.start,
         children: [
-          title != null ? Text(title!, style: TextStyle(color: Colors.grey.shade600),) : const SizedBox.shrink(),
-          title != null ? const SizedBox(height: 4,) : const SizedBox.shrink(),
+          title != null
+              ? Text(
+                  title!,
+                  style: TextStyle(color: Colors.grey.shade600),
+                )
+              : const SizedBox.shrink(),
+          title != null
+              ? const SizedBox(
+                  height: 4,
+                )
+              : const SizedBox.shrink(),
           ClipRRect(
             borderRadius: BorderRadius.circular(10.0),
             child: Material(