+++ lib/settings_switch_tile/settings_switch_tile_ios.dart
@@ -11,18 +11,17 @@ class SettingsSwitchTileIos extends StatefulWidget {
   final Color? thumbColor;
   final Color? trackColor;
 
-
-  const SettingsSwitchTileIos({
-    Key? key,
-    required this.title,
-    required this.onChanged,
-    required this.value,
-    this.subTitle = "",
-    this.icon,
-    this.activeColor,
-    this.thumbColor,
-    this.trackColor
-  }) : super(key: key);
+  const SettingsSwitchTileIos(
+      {Key? key,
+      required this.title,
+      required this.onChanged,
+      required this.value,
+      this.subTitle = "",
+      this.icon,
+      this.activeColor,
+      this.thumbColor,
+      this.trackColor})
+      : super(key: key);
 
   @override
   State<SettingsSwitchTileIos> createState() => _SettingsSwitchTileIosState();
@@ -36,17 +35,25 @@ class _SettingsSwitchTileIosState extends State<SettingsSwitchTileIos> {
       children: [
         ListTile(
           dense: true,
-          title: Text(widget.title, textAlign: TextAlign.start,),
+          title: Text(
+            widget.title,
+            textAlign: TextAlign.start,
+          ),
           leading: widget.icon,
           trailing: CupertinoSwitch(
-            onChanged: (bool value) {widget.onChanged(value);},
+            onChanged: (bool value) {
+              widget.onChanged(value);
+            },
             value: widget.value,
             activeColor: widget.activeColor,
             thumbColor: widget.thumbColor,
             trackColor: widget.trackColor,
           ),
         ),
-        const Divider(indent: 50, height: 0,),
+        const Divider(
+          indent: 50,
+          height: 0,
+        ),
       ],
     );
   }